+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 7.2.0-dev
+Version 7.2.0
 - deprecations triggered in a test method are now reported as warnings
 - it is now possible to also set background color with ConsoleColors::color()
 - added assertion assertTriggersDeprecation